+ Rework stats + active spell cache + Spell Stats Calculation et plein de bonnes choses

+++ Assets/_System/ECS/Components/Stats/Health.cs
@@ -1,7 +1,6 @@
 using Unity.Entities;
 
-
 public struct Health : IComponentData
 {
-    public float Value;
+    public int Value;
 }
